Predict whether a person has cancer or not
Predicting whether a person has cancer or not, based on information about the patient on Healthcare dataset using Machine Learning algorithms.
In this article, I am using Healthcare dataset as "gene_expression.csv".

In this article, we are going to predict whether a person has cancer or not, based on information about the patient such as Gene One, Gene Two. on Healthcare dataset using Machine Learning algorithms. The algorithms included KNN Algorithm.