Predicting the species of a penguin
Predicting the species of a penguin based on physical attributes on penguins_size.csv dataset using Machine Learning algorithms.
In this article, we are going to predict the species of a penguin based on physical attributes on penguins_size.csv dataset using Machine Learning algorithms. The algorithms included Decision Tree Algorithm.