Predicting the species of a penguin
Predicting the species of a penguin based on physical attributes on penguins_size.csv dataset using Machine Learning algorithms.
In this article, we will use the same dataset "penguins_size.csv", which we have used in previous in Decision Tree Algorithm. By using the same dataset, we can compare the Random Forest classifier with Decision tree. Here we are going to predict the species of a penguin based on physical attributes on penguins_size.csv dataset using Machine Learning algorithms. The algorithms included Random Forest Algorithm.