Mall Customers Segmentation
Mall Customers Dataset.  So, we have a dataset of "Mall_Customers.csv", which is the data of customers who visit the mall and spend there. In this article, we are going to segment these customers based on the details given, by using Machine Learning algorithms. The algorithms included K-means Clustering Algorithm.