Association Rule Mining - Grocery Store Data Set
In this article, we have a "GroceryStoreDataSet.csv" dataset, here we are going to identify strong relations discovered in products using some measures such as confidence or lift by using Apriori algorithm. Here we are identifying the products associations which having min_support value with a threshold value of 20% and 60% minimum confidence value. In other words, when product X is purchased, we can say that the purchase of product Y is 60% or more.

Goal is to identify strong relations discovered in products using some measures such as confidence or lift and identifying the association item. For the given dataset, apply the apriori algorithm to discover strong association rules. Assume that the min_support=20% and min_confidence=60%. Generate association rules from the frequent itemsets, calculate the confidence of each rule and identify all the strong association rules.

I have a dataset that I downloaded on https://www.kaggle.com/shazadudwadia/supermarket.